requirements.txt
openpyxl
pandas
azure-ai-anomalydetector
azure-core 
python-decouple

